se agrega categorias

+++ PedidosBarrio.Application/Commands/CreateProducto/CreateProductoCommand.cs
@@ -6,11 +6,11 @@ namespace PedidosBarrio.Application.Commands.CreateProducto
     public class CreateProductoCommand : IRequest<ProductoDto>
     {
         public Guid EmpresaID { get; set; }
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
 
-        public CreateProductoCommand(Guid empresaID, short categoriaID, string nombre, string descripcion)
+        public CreateProductoCommand(Guid empresaID, int categoriaID, string nombre, string descripcion)
         {
             EmpresaID = empresaID;
             CategoriaID = categoriaID;
