+++ PedidosBarrio.Application/Commands/UpdateProducto/UpdateProductoCommand.cs
@@ -7,11 +7,11 @@ namespace PedidosBarrio.Application.Commands.UpdateProducto
     {
         public int ProductoID { get; set; }
         public Guid EmpresaID { get; set; }
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
 
-        public UpdateProductoCommand(int productoID, Guid empresaID, short categoriaID, string nombre, string descripcion)
+        public UpdateProductoCommand(int productoID, Guid empresaID, int categoriaID, string nombre, string descripcion)
         {
             ProductoID = productoID;
             EmpresaID = empresaID;
