+++ PedidosBarrio.Application/DTOs/CreateProductoDto.cs
@@ -3,7 +3,7 @@ namespace PedidosBarrio.Application.DTOs
     public class CreateProductoDto
     {
         public Guid EmpresaID { get; set; }
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
     }
