+++ PedidosBarrio.Application/DTOs/NegocioDetalleDto.cs
@@ -21,7 +21,7 @@ namespace PedidosBarrio.Application.DTOs
     /// </summary>
     public class CategoriaDetalleDto
     {
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Descripcion { get; set; }
         public string Codigo { get; set; }
         public bool Mostrar { get; set; }
@@ -34,7 +34,7 @@ namespace PedidosBarrio.Application.DTOs
     {
         public int ProductoID { get; set; }
         public Guid EmpresaID { get; set; }
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public decimal Precio { get; set; }
