+++ PedidosBarrio.Domain/Entities/Producto.cs
@@ -4,7 +4,7 @@ namespace PedidosBarrio.Domain.Entities
     {
         public int ProductoID { get; set; }
         public Guid EmpresaID { get; set; }
-        public short CategoriaID { get; set; }
+        public int CategoriaID { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public DateTime FechaCreacion { get; set; }
